pyproject.toml
[tool.poetry]
name = "tsd-file-api"
version = "2.10.4"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
jwcrypto = "^1.3.1"
tornado = "^6.2"
psycopg2-binary = "^2.9.3"
libnacl = "^1.8.0"
pysquril = {url = "https://github.com/unioslo/pysquril/archive/refs/tags/v0.1.4.tar.gz"}
tsd-api-client = "^3.5.2"
SQLAlchemy = "^1.4.40"
pretty-bad-protocol = "^3.1.1"
click = "^8.1.3"
termcolor = "^1.1.0"
python-magic = "^0.4.27"
pika = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
